[chore][receiver/azuremonitorreceiver] move client options tests in a dedicated file
